refactor(Val::GUI): add burndown curve to work view
