+++ src/ramstk/views/gtk3/validation/__init__.py
@@ -11,4 +11,5 @@
 # from .listview import FailureDefinition as lvwFailureDefinition
 # from .listview import UsageProfile as lvwUsageProfile
 from .moduleview import ModuleView as mvwValidation
+from .workview import BurndownCurve as wvwBurndownCurve
 from .workview import GeneralData as wvwValidationGD
